Add base run tests file and documentation about ruby tests (#34)
